+++ crates/wdl-lsp/tests/notifications/main.rs
@@ -1,6 +1,5 @@
 //! Integration tests for the LSP notifications.
 
-mod did_change;
 mod did_change_configuration;
 mod did_change_watched_files;
 mod did_change_workspace_folders;